+++ theme/wrapRootElement.js
@@ -1,14 +1,12 @@
 /** @jsx jsx */
 import { jsx } from 'theme-ui'
 import { MDXProvider } from '@mdx-js/react'
-import { Provider as ReduxProvider } from 'react-redux'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { Themed } from '@theme-ui/mdx'
 
 import Emoji from './src/shortcodes/emoji'
 import Note from './src/shortcodes/Note'
 import RootWrapper from './src/components/root-wrapper'
-import store from './src/store'
 import theme from '@chronogrove/ui/theme'
 import YouTube from './src/shortcodes/youtube'
 import { ChronogroveThemeProvider } from '@chronogrove/ui'
@@ -37,13 +35,11 @@ const components = {
 
 const WrapRootElement = ({ element }) => (
   <QueryClientProvider client={queryClient}>
-    <ReduxProvider store={store}>
-      <ChronogroveThemeProvider theme={theme}>
-        <MDXProvider components={components}>
-          <RootWrapper>{element}</RootWrapper>
-        </MDXProvider>
-      </ChronogroveThemeProvider>
-    </ReduxProvider>
+    <ChronogroveThemeProvider theme={theme}>
+      <MDXProvider components={components}>
+        <RootWrapper>{element}</RootWrapper>
+      </MDXProvider>
+    </ChronogroveThemeProvider>
   </QueryClientProvider>
 )